Backend/requirements.txt
# Core dependencies for the Agricultural AI Assistant
# Image Processing
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.0.1
scikit-learn==1.3.0

# Speech Recognition and Text-to-Speech
SpeechRecognition==3.10.0
pydub==0.25.1
pyttsx3==2.90
gtts==2.3.2
pygame==2.5.2

# Advanced AI/ML (optional but recommended)
# Uncomment these for enhanced features:
# torch==2.0.1
# torchvision==0.15.2
# tensorflow==2.13.0
# openai-whisper==20230918
# edge-tts==6.1.9

# Additional audio processing
# pyaudio==0.2.11  # May require system-level installation

# For RAG system (if using)
# sentence-transformers==2.2.2
# chromadb==0.4.15
# langchain==0.0.308

# Utility packages
python-dotenv==1.0.0
requests==2.31.0